backend/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
facenet-pytorch==2.5.3
filelock==3.13.4
Flask==3.0.3
fsspec==2024.3.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pgvector==0.2.5
pillow==10.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
requests==2.31.0
sympy==1.12
torch==2.3.0
torchvision==0.18.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
